+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/controller/RecipeController.java
@@ -1,17 +1,17 @@
 package com.mars.app.domain.recipe.controller;
 
-import com.mars.app.domain.recipe.message.RecipeLikeMessagePublisher;
+import com.mars.app.domain.recipe.message.like.RecipeLikeMessagePublisher;
 import com.mars.app.domain.recipe.event.RecipeLikeSseService;
 import com.mars.common.dto.ResponseDto;
 import com.mars.app.aop.auth.AuthenticatedUser;
 import com.mars.app.component.auth.AuthenticationHolder;
 import com.mars.common.dto.page.PageRequestVO;
 import com.mars.app.domain.recipe.dto.RecipeEsResponseDto;
-import com.mars.app.domain.recipe.dto.RecipeLikeResponseDto;
+import com.mars.app.domain.recipe.dto.like.RecipeLikeResponseDto;
 import com.mars.app.domain.recipe.dto.RecipeResponseDto;
 import com.mars.app.domain.recipe.service.RecipeEsService;
 import com.mars.app.domain.recipe.service.RecipeEsSynchronizerService;
-import com.mars.app.domain.recipe.service.RecipeLikeService;
+import com.mars.app.domain.recipe.service.like.RecipeLikeService;
 import com.mars.app.domain.recipe.service.RecipeService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
