+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/dto/favorite/RecipeFavoriteRequestDto.java
@@ -1,4 +1,4 @@
-package com.mars.app.domain.favorite.recipe.dto;
+package com.mars.app.domain.recipe.dto.favorite;
 
 public record RecipeFavoriteRequestDto(
     Long recipeId
