+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/dto/like/RecipeLikeRequestDto.java
@@ -1,4 +1,4 @@
-package com.mars.app.domain.recipe.dto;
+package com.mars.app.domain.recipe.dto.like;
 
 public record RecipeLikeRequestDto(
     Long recipeId
