+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/message/like/RecipeLikeMessagePublisher.java
@@ -1,8 +1,8 @@
-package com.mars.app.domain.recipe.message;
+package com.mars.app.domain.recipe.message.like;
 
 import com.mars.app.config.rabbitmq.impl.RecipeLikeRabbitConfig;
-import com.mars.app.domain.recipe.dto.RecipeLikeMessageDto;
-import com.mars.app.domain.recipe.dto.RecipeLikeResponseDto;
+import com.mars.app.domain.recipe.dto.like.RecipeLikeMessageDto;
+import com.mars.app.domain.recipe.dto.like.RecipeLikeResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.amqp.core.TopicExchange;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
