+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/repository/like/RecipeLikeRepository.java
@@ -1,4 +1,4 @@
-package com.mars.app.domain.recipe.repository;
+package com.mars.app.domain.recipe.repository.like;
 
 import com.mars.common.model.recipe.Recipe;
 import com.mars.common.model.recipe.RecipeLike;
