+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/service/like/RecipeLikeService.java
@@ -1,6 +1,6 @@
-package com.mars.app.domain.recipe.service;
+package com.mars.app.domain.recipe.service.like;
 
-import com.mars.app.domain.recipe.repository.RecipeLikeRepository;
+import com.mars.app.domain.recipe.repository.like.RecipeLikeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
